Parsing CSV files
Playing with the goalie and players csv files, thought i had to download fresh data for this, but  maybe the optimizer is doing that work for me?  pulling from DK or FD depedning on which one you choose during first run.  I hope this only updates my REPO, new to all of this, learning as i go!!!